cleanup layout
