--- examples/betterauth/env.d.ts
@@ -1,6 +1,0 @@
-declare namespace NodeJS {
-  interface ProcessEnv {
-    GITHUB_CLIENT_ID?: string;
-    GITHUB_CLIENT_SECRET?: string;
-  }
-}
