+++ examples/betterauth/src/app/(app)/settings/page.tsx
@@ -1,5 +1,4 @@
-import SettingsForm from "@/components/forms/Settings";
-import { Navbar } from "@/components/navbar";
+import { UserSettings } from "@/components/user-settings";
 import type { Metadata } from "next";
 
 export const metadata: Metadata = {
@@ -8,9 +7,8 @@ export const metadata: Metadata = {
 
 export default function SettingsPage() {
   return (
-    <>
-      <Navbar />
-      <SettingsForm />;
-    </>
+    <div className="max-w-screen-md w-full mx-auto px-4">
+      <UserSettings />
+    </div>
   );
 }
